+++ index.html
@@ -245,7 +245,6 @@
   </div>
 </template>
 
-<script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
-<script src="./js/main.js" defer></script>
+<script src="./js/main.js" type="module" defer></script>
 </body>
 </html>
