+++ js/util.js
@@ -1,0 +1,45 @@
+export {getRandomInt, getRandomFloat, getRandomArray, getRandomArrayWithUniqueItems};
+
+const getRandomInt = (min, max) => {
+  if (min < 0 || max < 0 || min > max) {
+    return undefined;
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+
+const getRandomFloat = (min, max, symbolsNumber) => {
+  if (min < 0 || max < 0 || min > max) {
+    return undefined;
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return (Math.random() * (max - min) + min).toFixed(symbolsNumber);
+}
+
+
+const getRandomArray = (array) => {
+  const randomArray = [];
+  for (let i = 0; i < getRandomInt(0, array.length); i++) {
+    randomArray.push(array[getRandomInt(0, getRandomInt.length - 1)]);
+  }
+  return randomArray;
+};
+
+
+const getRandomArrayWithUniqueItems = (array) => {
+  const arrayLength = getRandomInt(0, array.length);
+  const resultRandomArrayWithUniqueItems = [];
+  while (resultRandomArrayWithUniqueItems.length < arrayLength) {
+    const maxIndex = array.length - 1;
+    const randomIndex = getRandomInt(0, maxIndex);
+    const randomItem = array[randomIndex];
+
+    if (!resultRandomArrayWithUniqueItems.includes(randomItem)) {
+      resultRandomArrayWithUniqueItems.push(randomItem);
+    }
+  }
+  return resultRandomArrayWithUniqueItems;
+}
